refactor: move prompt related functionalities to separate file
